refactor: link editor
